refactor(program): extract common startup for lambda entrypoint
