+++ ChauMundo/LambdaEntryPoint.cs
@@ -1,0 +1,13 @@
+using Amazon.Lambda.AspNetCoreServer;
+
+namespace ChauMundo
+{
+    public class LambdaEntryPoint : APIGatewayProxyFunction
+    {
+        protected override void Init(IWebHostBuilder builder)
+        {
+            builder
+                .UseStartup<Startup>();
+        }
+    }
+}
