+++ ChauMundo/Startup.cs
@@ -1,0 +1,44 @@
+using ChauMundo.Printers;
+
+namespace ChauMundo
+{
+    public class Startup
+    {
+        public Startup(IConfiguration configuration)
+        {
+            Configuration = configuration;
+        }
+
+        public IConfiguration Configuration { get; }
+
+        public void ConfigureServices(IServiceCollection services)
+        {
+            services.AddEndpointsApiExplorer();
+            services.AddSwaggerGen();
+        }
+
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+        {
+            if (env.IsDevelopment())
+            {
+                app.UseDeveloperExceptionPage();
+                app.UseSwagger();
+                app.UseSwaggerUI();
+            }
+
+            app.UseHttpsRedirection();
+
+            app.UseRouting();
+
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapGet("/hello", () =>
+                {
+                    string? message = Configuration["Message"];
+                    Program.PrintMessage(new DebugPrinter(), message ?? "Default message");
+                    return Results.Ok("Message printed to debug output");
+                });
+            });
+        }
+    }
+}
